fix: go directly to search location screen when adding favorite
Fixes #674
