+++ src/components/sections/label-item.tsx
@@ -4,19 +4,11 @@ import {StyleSheet} from '../../theme';
 import ThemeText from '../text';
 import {useSectionItem, SectionItem, useSectionStyle} from './section-utils';
 
-export type HeaderItemProps = SectionItem<{
+export type LabelItemProps = SectionItem<{
   text: string;
   prefix?: string;
-  subtitle?: string;
-  mode?: 'heading' | 'subheading';
 }>;
-export default function HeaderItem({
-  text,
-  prefix,
-  subtitle,
-  mode = 'heading',
-  ...props
-}: HeaderItemProps) {
+export default function LabelItem({text, prefix, ...props}: LabelItemProps) {
   const {contentContainer, topContainer} = useSectionItem(props);
   const style = useSectionStyle();
   const headerStyle = useHeaderStyle();
@@ -25,23 +17,8 @@ export default function HeaderItem({
     <View style={topContainer}>
       <View style={style.spaceBetween}>
         {prefix && <ThemeText style={headerStyle.prefix}>{prefix}</ThemeText>}
-        <ThemeText
-          style={contentContainer}
-          color={mode === 'heading' ? 'primary' : 'faded'}
-          type={mode === 'heading' ? 'paragraphHeadline' : 'lead'}
-        >
-          {text}
-        </ThemeText>
+        <ThemeText style={contentContainer}>{text}</ThemeText>
       </View>
-      {subtitle && (
-        <ThemeText
-          style={contentContainer}
-          color="faded"
-          type={mode === 'heading' ? 'lead' : 'label'}
-        >
-          {subtitle}
-        </ThemeText>
-      )}
     </View>
   );
 }
