+++ src/location-search/LocationSearch.tsx
@@ -1,6 +1,11 @@
 import {RouteProp, useIsFocused} from '@react-navigation/native';
 import React, {useEffect, useRef, useState} from 'react';
-import {Keyboard, TextInput as InternalTextInput, View} from 'react-native';
+import {
+  Keyboard,
+  SegmentedControlIOSComponent,
+  TextInput as InternalTextInput,
+  View,
+} from 'react-native';
 import {ScrollView} from 'react-native-gesture-handler';
 import {TRONDHEIM_CENTRAL_STATION} from '../api/geocoder';
 import {ErrorType} from '../api/utils';
@@ -23,15 +28,15 @@ import {RootStackParamList} from '../navigation';
 import FullScreenHeader from '../ScreenHeader/full-header';
 import {useSearchHistory} from '../search-history';
 import {StyleSheet} from '../theme';
+import {
+  LocationSearchTexts,
+  TranslateFunction,
+  useTranslation,
+} from '../translations/';
 import {LocationSearchNavigationProp} from './';
 import LocationResults from './LocationResults';
 import {LocationSearchResult} from './types';
 import useDebounce from './useDebounce';
-import {
-  useTranslation,
-  LocationSearchTexts,
-  TranslateFunction,
-} from '../translations/';
 export type Props = {
   navigation: LocationSearchNavigationProp;
   route: RouteProp<RootStackParamList, 'LocationSearch'>;
@@ -57,31 +62,11 @@ const LocationSearch: React.FC<Props> = ({
     },
   },
 }) => {
-  const styles = useThemeStyles();
-  const {history, addSearchEntry} = useSearchHistory();
-  const {favorites} = useFavorites();
-
-  const [text, setText] = useState<string>(initialLocation?.name ?? '');
-  const debouncedText = useDebounce(text, 200);
   const {t} = useTranslation();
-
-  const [errorMessage, setErrorMessage] = useState<string>('');
-  const previousLocations = filterPreviousLocations(
-    debouncedText,
-    history,
-    favorites,
-  );
-
+  const styles = useThemeStyles();
   const {location: geolocation} = useGeolocationState();
 
-  const {locations, error} =
-    useGeocoder(debouncedText, geolocation?.coords ?? null) ?? [];
-  const filteredLocations = filterCurrentLocation(locations, previousLocations);
-
   const onSelect = (location: LocationWithMetadata) => {
-    if (location.resultType === 'search') {
-      addSearchEntry(location);
-    }
     navigation.navigate(callerRouteName as any, {
       [callerRouteParam]: location,
     });
@@ -105,7 +90,88 @@ const LocationSearch: React.FC<Props> = ({
     });
   };
 
+  return (
+    <View style={styles.container}>
+      <FullScreenHeader
+        title={t(LocationSearchTexts.header.title)}
+        leftButton={{
+          onPress: () => navigation.goBack(),
+          accessible: true,
+          accessibilityRole: 'button',
+          accessibilityLabel: t(
+            LocationSearchTexts.header.leftButton.a11yLabel,
+          ),
+          icon: <ThemeIcon svg={Close} />,
+        }}
+      />
+
+      <LocationSearchContent
+        onSelect={onSelect}
+        onMapSelection={onMapSelection}
+        label={label}
+        favoriteChipTypes={favoriteChipTypes}
+        placeholder={t(LocationSearchTexts.searchField.placeholder)}
+        defaultText={initialLocation?.name}
+      />
+    </View>
+  );
+};
+
+type LocationSearchContentProps = {
+  label: string;
+  placeholder: string;
+  favoriteChipTypes?: ChipTypeGroup[];
+  defaultText?: string;
+  onSelect(location: LocationWithMetadata): void;
+  onMapSelection?(): void;
+  onlyAtbVenues?: boolean;
+  includeHistory?: boolean;
+};
+
+export function LocationSearchContent({
+  label,
+  placeholder,
+  favoriteChipTypes,
+  defaultText,
+  onSelect,
+  onMapSelection,
+  onlyAtbVenues = false,
+  includeHistory = true,
+}: LocationSearchContentProps) {
+  const styles = useThemeStyles();
+  const {favorites} = useFavorites();
+  const {history, addSearchEntry} = useSearchHistory();
+  const {t} = useTranslation();
+
+  const [text, setText] = useState<string>(defaultText ?? '');
+  const debouncedText = useDebounce(text, 200);
+
+  const [errorMessage, setErrorMessage] = useState<string>('');
+  const previousLocations = filterPreviousLocations(
+    debouncedText,
+    history,
+    favorites,
+    onlyAtbVenues,
+  );
+
+  const {location: geolocation} = useGeolocationState();
+
+  const {locations, error} = useGeocoder(
+    debouncedText,
+    geolocation?.coords ?? null,
+    onlyAtbVenues,
+  );
+
+  const filteredLocations = filterCurrentLocation(
+    locations,
+    includeHistory ? previousLocations : [],
+  );
+
   const onSearchSelect = (searchResult: LocationSearchResult) => {
+    if (!searchResult.favoriteInfo) {
+      addSearchEntry(searchResult.location);
+    }
+
     if (!searchResult.favoriteInfo) {
       return onSelect({...searchResult.location, resultType: 'search'});
     }
@@ -141,23 +207,10 @@ const LocationSearch: React.FC<Props> = ({
 
   const hasPreviousResults = !!previousLocations.length;
   const hasResults = !!filteredLocations.length;
-  const hasAnyResult = hasResults || hasPreviousResults;
+  const hasAnyResult = hasResults || (includeHistory && hasPreviousResults);
 
   return (
-    <View style={styles.container}>
-      <FullScreenHeader
-        title={t(LocationSearchTexts.header.title)}
-        leftButton={{
-          onPress: () => navigation.goBack(),
-          accessible: true,
-          accessibilityRole: 'button',
-          accessibilityLabel: t(
-            LocationSearchTexts.header.leftButton.a11yLabel,
-          ),
-          icon: <ThemeIcon svg={Close} />,
-        }}
-      />
-
+    <>
       <View style={styles.header}>
         <ScreenReaderAnnouncement message={errorMessage} />
 
@@ -170,7 +223,7 @@ const LocationSearch: React.FC<Props> = ({
             onChangeText={setText}
             showClear={Boolean(text?.length)}
             onClear={() => setText('')}
-            placeholder={t(LocationSearchTexts.searchField.placeholder)}
+            placeholder={placeholder}
             autoCorrect={false}
             autoCompleteType="off"
           />
@@ -195,12 +248,12 @@ const LocationSearch: React.FC<Props> = ({
 
       {hasAnyResult ? (
         <ScrollView
-          style={styles.scroll}
+          style={styles.fullFlex}
           contentContainerStyle={styles.contentBlock}
           keyboardShouldPersistTaps="handled"
           onScrollBeginDrag={() => Keyboard.dismiss()}
         >
-          {hasPreviousResults && (
+          {includeHistory && hasPreviousResults && (
             <LocationResults
               title={t(LocationSearchTexts.results.previousResults.heading)}
               locations={previousLocations}
@@ -229,9 +282,9 @@ const LocationSearch: React.FC<Props> = ({
           </View>
         )
       )}
-    </View>
+    </>
   );
-};
+}
 
 function translateErrorType(
   errorType: ErrorType,
@@ -250,11 +303,16 @@ const filterPreviousLocations = (
   searchText: string,
   previousLocations: Location[],
   favorites?: UserFavorites,
+  onlyAtbVenues: boolean = false,
 ): LocationSearchResult[] => {
   const mappedHistory: LocationSearchResult[] =
-    previousLocations?.map((location) => ({
-      location,
-    })) ?? [];
+    previousLocations
+      ?.map((location) => ({
+        location,
+      }))
+      .filter(
+        (location) => !onlyAtbVenues || location.location.layer == 'venue',
+      ) ?? [];
 
   if (!searchText) {
     return mappedHistory;
@@ -265,7 +323,8 @@ const filterPreviousLocations = (
   const filteredFavorites: LocationSearchResult[] = (favorites ?? [])
     .filter(
       (favorite) =>
-        matchText(favorite.location?.name) || matchText(favorite.name),
+        (matchText(favorite.location?.name) || matchText(favorite.name)) &&
+        (!onlyAtbVenues || favorite.location.layer == 'venue'),
     )
     .map(({location, ...favoriteInfo}) => ({
       location,
@@ -305,7 +364,7 @@ const useThemeStyles = StyleSheet.createThemeHook((theme) => ({
   contentBlock: {
     paddingHorizontal: theme.spacings.medium,
   },
-  scroll: {
+  fullFlex: {
     flex: 1,
   },
 }));
