+++ src/sdk.ts
@@ -14,6 +14,7 @@ import {
   ServiceJourney,
   StopPlace,
   StopPlaceDetails,
+  TransportMode,
   TransportSubmode,
   ValidityPeriod,
 } from '@entur/sdk';
@@ -190,10 +191,19 @@ export type DeparturesRealtimeData = {
   [quayId: string]: DepartureRealtimeData;
 };
 
-export type FavoriteDeparture = {
+export type FavoriteDepartureId = {
   stopId: string;
   lineName: string;
   lineId: string;
+};
+
+export type FavoriteDeparture = FavoriteDepartureId & {
+  lineLineNumber?: string;
+  lineTransportationMode?: TransportMode;
+  lineTransportationSubMode?: TransportSubmode;
+  quayName: string;
+  quayId: string;
+  quayPublicCode?: string;
 };
 
 export type NextCursorData = {
